move offsetFn
